# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  git_neolidy_top_neo_flowx_internal_application_notification.CreateNotificationRequest:
    properties:
      category:
        type: string
      channel:
        type: string
      content:
        type: string
      extra:
        $ref: '#/definitions/git_neolidy_top_neo_flowx_internal_domain_base.JSON'
      receiver_id:
        type: string
      ref_id:
        type: string
      ref_type:
        type: string
      title:
        type: string
      type:
        type: string
    required:
    - receiver_id
    - title
    - type
    type: object
  git_neolidy_top_neo_flowx_internal_application_notification.CreatePreferenceRequest:
    properties:
      channel:
        type: string
      enabled:
        type: boolean
      mute_until:
        type: string
      type:
        type: string
      user_id:
        type: string
    required:
    - channel
    - type
    - user_id
    type: object
  git_neolidy_top_neo_flowx_internal_application_notification.CreateTemplateRequest:
    properties:
      channel:
        type: string
      code:
        type: string
      content_tpl:
        type: string
      description:
        type: string
      name:
        type: string
      title_tpl:
        type: string
      type:
        type: string
      variables:
        $ref: '#/definitions/git_neolidy_top_neo_flowx_internal_domain_base.JSON'
    required:
    - channel
    - code
    - content_tpl
    - name
    - title_tpl
    - type
    type: object
...
  git_neolidy_top_neo_flowx_internal_application_notification.UpdatePreferenceRequest:
    properties:
      enabled:
        type: boolean
      mute_until:
        type: string
    type: object
...
  git_neolidy_top_neo_flowx_pkg_response.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: neolidy@outlook.com
    name: jiangfire
    url: https://github.com/jiangfire/flowx
  description: 企业级智能工具治理与流程编排平台 API 文档
  license:
    name: MIT
    url: https://github.com/jiangfire/flowx/blob/main/LICENSE
  termsOfService: https://github.com/jiangfire/flowx
  title: FlowX API
  version: "1.0"
paths:
...
  /notification-preferences:
    get:
      consumes:
      - application/json
      description: 分页查询通知偏好列表，支持按用户、类型、渠道筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 用户ID
        in: query
        name: user_id
        type: string
      - description: 偏好类型
        in: query
        name: type
        type: string
      - description: 通知渠道
        in: query
        name: channel
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 通知偏好列表
      tags:
      - 通知系统
    post:
      consumes:
      - application/json
      description: 创建新的通知偏好设置
      parameters:
      - description: 创建通知偏好请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/git_neolidy_top_neo_flowx_internal_application_notification.CreatePreferenceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 创建通知偏好
      tags:
      - 通知系统
  /notification-preferences/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除通知偏好设置
      parameters:
      - description: 通知偏好ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 删除通知偏好
      tags:
      - 通知系统
    put:
      consumes:
      - application/json
      description: 根据ID更新通知偏好设置
      parameters:
      - description: 通知偏好ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新通知偏好请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/git_neolidy_top_neo_flowx_internal_application_notification.UpdatePreferenceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 更新通知偏好
      tags:
      - 通知系统
  /notification-templates:
    get:
      consumes:
      - application/json
      description: 分页查询通知模板列表，支持按类型、状态筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 模板类型
        in: query
        name: type
        type: string
      - description: 模板状态
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 通知模板列表
      tags:
      - 通知系统
    post:
      consumes:
      - application/json
      description: 创建新的通知模板
      parameters:
      - description: 创建通知模板请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/git_neolidy_top_neo_flowx_internal_application_notification.CreateTemplateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 创建通知模板
      tags:
      - 通知系统
...
  /notifications:
    get:
      consumes:
      - application/json
      description: 分页查询通知列表，支持按类型、分类、已读状态筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 通知类型
        in: query
        name: type
        type: string
      - description: 通知分类
        in: query
        name: category
        type: string
      - description: 已读状态
        in: query
        name: is_read
        type: string
      - description: 通知状态
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 通知列表
      tags:
      - 通知系统
    post:
      consumes:
      - application/json
      description: 创建新的通知
      parameters:
      - description: 创建通知请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/git_neolidy_top_neo_flowx_internal_application_notification.CreateNotificationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 创建通知
      tags:
      - 通知系统
...
  /notifications/{id}/read:
    put:
      consumes:
      - application/json
      description: 将指定通知标记为已读
      parameters:
      - description: 通知ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 标记通知为已读
      tags:
      - 通知系统
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer Token，格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"